requirements.txt
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.31.1
boto3==1.34.73
botocore==1.31.85
Flask==3.0.2
virtualenv==20.25.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.0
responses==0.18.0
langchain-text-splitters==0.0.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.92
libclang==14.0.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.6.1
waitress==3.0.0